Rename to DeepLinkKit
