+++ .slather.yml
@@ -1,3 +1,3 @@
 coverage_service: coveralls
-xcodeproj: DeepLinkSDK.xcodeproj
-source_directory: DeepLinkSDK
+xcodeproj: DeepLinkKit.xcodeproj
+source_directory: DeepLinkKit
