+++ DeepLinkKit.podspec
@@ -1,5 +1,5 @@
 Pod::Spec.new do |s|
-  s.name             = "DeepLinkSDK"
+  s.name             = "DeepLinkKit"
   s.version          = "0.2.6"
   s.summary          = "A splendid route-matching, block-based way to handle your deep links."
   s.description      = <<-DESC
@@ -8,11 +8,11 @@ Pod::Spec.new do |s|
   s.homepage         = "http://www.usebutton.com/sdk/deep-links"
   s.license          = 'MIT'
   s.author           = 'Button, Inc.'
-  s.source           = { :git => "https://github.com/usebutton/ios-deeplink-sdk.git", :tag => s.version.to_s }
+  s.source           = { :git => "https://github.com/usebutton/DeepLinkKit.git", :tag => s.version.to_s }
   s.social_media_url = 'https://twitter.com/buttondev'
 
   s.platform         = :ios, '7.0'
   s.requires_arc     = true
 
-  s.source_files     = 'DeepLinkSDK/**/*.{h,m}'
+  s.source_files     = 'DeepLinkKit/**/*.{h,m}'
 end
