+++ DeepLinkKit/AppLinks.h
@@ -1,3 +1,3 @@
-#import "DeepLinkSDK.h"
+#import "DeepLinkKit.h"
 #import "DPLDeepLink+AppLinks.h"
 #import "DPLMutableDeepLink+AppLinks.h"
