+++ Rakefile
@@ -1,11 +1,11 @@
 desc "Runs the specs"
 task :spec do
-  sh "xcodebuild test -workspace 'DeepLinkSDK.xcworkspace' -scheme 'DeepLinkSDK' -destination 'platform=iOS Simulator,name=iPhone 6,OS=8.1' | xcpretty -c"
+  sh "xcodebuild test -workspace 'DeepLinkKit.xcworkspace' -scheme 'DeepLinkKit' -destination 'platform=iOS Simulator,name=iPhone 6,OS=8.1' | xcpretty -c"
 end
 
 desc "Synchronizes Xcode project folder with Xcode groups"
 task :sync do
-  sh "synx -e '/Pod Metadata' -e '/Pods' DeepLinkSDK.xcodeproj"
+  sh "synx -e '/Pod Metadata' -e '/Pods' DeepLinkKit.xcodeproj"
 end
 
 desc "Runs pod install with --no-repo-update flag"
